Split vector plan construction by responsibility
R5.3 of the transform cleanup plan: vector/plan/mod.rs becomes build.rs
(placement, edges, effect orientation), fusion.rs (fused serial
groups), and producer_reachability.rs (bitset reachability, documented
as deliberately separate from the checker's reachability) behind an
unchanged facade keeping VerifiedVectorPlan and the build-error
taxonomy. Move only: pub(super) on sibling-crossed items and per-file
imports. R5 is complete.

transform 385, vector_mode 35/35, clippy clean.
