+++ porting/journal/2026-07-20.md
@@ -1,0 +1,27 @@
+# 2026-07-20
+
+## Split vector plan construction by responsibility (R5.3)
+
+Commit date: 2026-07-20
+Source commit: this commit
+
+`vector/plan/mod.rs` (1,403 production lines) becomes three files behind an
+unchanged facade:
+
+- `build.rs` (1,027) — `build_vector_plan(_with_lockstep)`, `PlacementState`,
+  ordinary dependency edges, effect-conflict orientation, and the producer's
+  type/rate conversions;
+- `fusion.rs` (556) — fused serial-group construction (sample closure,
+  dangerous delayed-read relations, conflicting effect users);
+- `producer_reachability.rs` (104) — `PlanReachability` bitset reachability,
+  header-documented as deliberately separate from the checker's
+  `verify::checker_reachability` (trust boundary, plan §4.6);
+- `mod.rs` (127) — facade, `VerifiedVectorPlan` wrapper, build-error
+  taxonomy, and the `#[cfg(test)]` `verified_vector_plan_for_test` helper.
+
+Move only: bodies unchanged apart from `pub(super)` on sibling-crossed
+items and per-file imports. With R5.1 (analysis) and R5.2 (verify), phase R5
+is complete: producer and checker reachability now live in visibly disjoint
+modules on both sides of the plan artifact. Validation: transform 385,
+`vector_mode` oracle 35/35, clippy `-D warnings` clean, fmt applied.
+
