+++ porting/journal/README.md
@@ -119,3 +119,4 @@ Files are listed chronologically (oldest day first). Inside each file, entries a
 - [2026-07-17](./2026-07-17.md) (20 entries)
 - [2026-07-18](./2026-07-18.md) (5 entries)
 - [2026-07-19](./2026-07-19.md) (18 entries)
+- [2026-07-20](./2026-07-20.md) (1 entry)
